+++ bare-metal/fedora-coreos/kubernetes/bootstrap.tf
@@ -1,6 +1,6 @@
 # Kubernetes assets (kubeconfig, manifests)
 module "bootstrap" {
-  source = "git::https://github.com/poseidon/terraform-render-bootstrap.git?ref=ae5449a9fbc439b5387582f5c45b5c0ba3859949"
+  source = "git::https://github.com/poseidon/terraform-render-bootstrap.git?ref=75fc91deb8ee740b8ce3297fbfbd451785bc9afb"
 
   cluster_name                    = var.cluster_name
   api_servers                     = [var.k8s_domain_name]