+++ bare-metal/flatcar-linux/kubernetes/bootstrap.tf
@@ -1,6 +1,6 @@
 # Kubernetes assets (kubeconfig, manifests)
 module "bootstrap" {
-  source = "git::https://github.com/poseidon/terraform-render-bootstrap.git?ref=75fc91deb8ee740b8ce3297fbfbd451785bc9afb"
+  source = "git::https://github.com/poseidon/terraform-render-bootstrap.git?ref=efd750d7a81228cdc47b81329fe5dbef1e6542b5"
 
   cluster_name                    = var.cluster_name
   api_servers                     = [var.k8s_domain_name]